+++ src/main/java/uk/gov/companieshouse/efs/web/security/CompanyAuthFilter.java
@@ -3,6 +3,11 @@ package uk.gov.companieshouse.efs.web.security;
 import uk.gov.companieshouse.auth.filter.AuthFilter;
 import uk.gov.companieshouse.efs.web.categorytemplates.service.api.CategoryTemplateService;
 import uk.gov.companieshouse.efs.web.formtemplates.service.api.FormTemplateService;
+import uk.gov.companieshouse.efs.web.security.validator.FormTemplateRequiredValidator;
+import uk.gov.companieshouse.efs.web.security.validator.HttpRequestRequiredValidator;
+import uk.gov.companieshouse.efs.web.security.validator.UserRequiredValidator;
+import uk.gov.companieshouse.efs.web.security.validator.Validator;
+import uk.gov.companieshouse.efs.web.security.validator.ValidatorResourceProvider;
 import uk.gov.companieshouse.efs.web.service.api.ApiClientService;
 import uk.gov.companieshouse.environment.EnvironmentReader;
 import uk.gov.companieshouse.session.Session;
