+++ src/main/java/uk/gov/companieshouse/efs/web/security/validator/ValidatorImpl.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.efs.web.security;
+package uk.gov.companieshouse.efs.web.security.validator;
 
 /**
  * Implementation of the validator interface.
