+++ src/test/java/uk/gov/companieshouse/efs/web/security/validator/ValidatorResourceProviderTest.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.efs.web.security;
+package uk.gov.companieshouse.efs.web.security.validator;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
